// ts/features/itwallet/identification/analytics/enum.ts
export enum ITW_IDENTIFICATION_SCREENVIEW_EVENTS {
  ITW_ID_METHOD = "ITW_ID_METHOD",
  ITW_SPID_IDP_SELECTION = "ITW_SPID_IDP_SELECTION",
  ITW_CIE_PIN_ENTER = "ITW_CIE_PIN_ENTER",
  ITW_CIE_NFC_ACTIVATION = "ITW_CIE_NFC_ACTIVATION",
  ITW_CIE_CARD_READING = "ITW_CIE_CARD_READING",
  ITW_CARD_READING_SUCCESS = "ITW_CARD_READING_SUCCESS",
  ITW_PIN_INFO_BOTTOMSHEET = "ITW_PIN_INFO_BOTTOMSHEET",
  ITW_CIE_INFO_BOTTOMSHEET = "ITW_CIE_INFO_BOTTOMSHEET",
  ITW_CIE_PIN_TUTORIAL_CIE = "ITW_CIE_PIN_TUTORIAL_CIE",
  ITW_CIE_PIN_TUTORIAL_PIN = "ITW_CIE_PIN_TUTORIAL_PIN",
  ITW_USER_WITHOUT_L3_BOTTOMSHEET = "ITW_USER_WITHOUT_L3_BOTTOMSHEET"
}

export enum ITW_ACTIONS_EVENTS {}